split file createMsgs.cpp into new files. Also renamed a few files and functions to more accurately reflect their function/content
