+++ src/privmsg.cpp
@@ -18,6 +18,21 @@ void	Server::userMessageToChannel(Channel channel, int sender_socket, std::strin
 	}
 }
 
+void		Server::channelMessage(Msg msg, int clientSocket, Client &client)
+{
+	int i = getChannelIndex(msg.parameters[0], channel_names);
+	if (i != -1)
+	{
+		//FORMAT -> :UserA!user@host PRIVMSG #general :Hello, world!
+		std::string message = ":" + client.getNickname() + "!" + client.getUsername() + "@" + this->_servHostName + " " + msg.command + " " + msg.parameters[0] + " "  + ":" +  msg.trailing_msg + "\r\n";
+		userMessageToChannel(channel_names[i], clientSocket, message);			
+	}
+	else
+	{
+		std::cout << "Channel to send message to not found" << std::endl;
+	}
+}
+
 void		Server::directMessage(Msg msg, int clientSocket, Client &client)
 {
 	int socket = getClientSocket(msg.parameters[0]);
@@ -31,21 +46,6 @@ void		Server::directMessage(Msg msg, int clientSocket, Client &client)
 		// FORMAT -> :client1!user@host PRIVMSG client2 :message		
 		std::string message = ":" + client.getNickname() + "!" + client.getUsername() + "@" + this->_servHostName + " " + msg.command + " " + msg.parameters[0] + " "  + ":" +  msg.trailing_msg + "\r\n";
 		send(socket, message.c_str(), message.size(), 0);
-	}
-}
-
-void		Server::channelMessage(Msg msg, int clientSocket, Client &client)
-{
-	int i = getChannelIndex(msg.parameters[0], channel_names);
-	if (i != -1)
-	{
-		//FORMAT -> :UserA!user@host PRIVMSG #general :Hello, world!
-		std::string message = ":" + client.getNickname() + "!" + client.getUsername() + "@" + this->_servHostName + " " + msg.command + " " + msg.parameters[0] + " "  + ":" +  msg.trailing_msg + "\r\n";
-		userMessageToChannel(channel_names[i], clientSocket, message);			
-	}
-	else
-	{
-		std::cout << "Channel to send message to not found" << std::endl;
 	}
 }
